Release version 1.01
- Completely build .exe file by Pyinstaller and Nuitka
Todo:
-Version 1.02: detect language
-Version 2.00: log history
