+++ .gitignore
@@ -1,0 +1,3 @@
+exe/build/
+exe/*.build/
+exe/*.dist/
